+++ src/main/java/com/myresume/librarymanagementsystem/gender/repository/GenderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenderRepository extends JpaRepository<Gender, Integer> {
+public interface GenderRepository extends JpaRepository<Gender, Long> {
 }
