+++ src/main/java/com/myresume/librarymanagementsystem/library/repository/LibraryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibraryRepository extends JpaRepository<Library, Integer> {
+public interface LibraryRepository extends JpaRepository<Library, Long> {
 }
